optimisation: cleanup print statistics
Change in memory:
Flash: -104 bytes
SRAM: 0 bytes

+++ Firmware/Marlin.h
@@ -312,8 +312,12 @@ extern ShortTimer usb_timer;
 extern bool processing_tcode;
 extern bool homing_flag;
 extern bool loading_flag;
-extern unsigned long total_filament_used;
-void save_statistics(unsigned long _total_filament_used, unsigned long _total_print_time);
+extern uint32_t total_filament_used; // mm/100 or 10um
+
+/// @brief Save print statistics to EEPROM
+/// @param _total_filament_used has unit mm/100 or 10um
+/// @param _total_print_time has unit minutes, for example 123 minutes
+void save_statistics(uint32_t _total_filament_used, uint32_t _total_print_time);
 extern uint8_t heating_status_counter;
 
 extern bool fan_state[2];
